@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-ts_proto. DO NOT EDIT.
 // versions:
 //   protoc-gen-ts_proto  v1.181.1
-//   protoc               v3.21.12
+//   protoc               v5.27.2
 // source: proto_chat.proto
 
 /* eslint-disable */
